Over9000
Takes one of several DBZ character's names as arguments:
Kakarot/Goku, Frieza, Vegeta. Sends a message in the discord chat
with a primitive, single-line saiyan text emoji and a quote with
the character's name

If no arguments are passed, the plugin defaults to Goku.
If an invalid argument is passed, an elipsis replaces the
would-be quote and name.

May add more characters and quotes in the future. Other features in
mind as well. Depends on the free time I have for tomfoolery and
shenanigans.